fix and organize unit tests
