+++ apiml-utility/src/main/java/org/zowe/apiml/product/logging/LogMessageTracker.java
@@ -27,7 +27,6 @@ import java.util.regex.Pattern;
  */
 public class LogMessageTracker {
     private final ListAppender<ILoggingEvent> logAppender = new ListAppender<>();
-    private final Logger log;
 
     /**
      * @param loggedClass Class that generates the logs to be searched.
@@ -42,16 +41,9 @@ public class LogMessageTracker {
      */
     public LogMessageTracker(String loggerName) {
         logAppender.setContext((LoggerContext) LoggerFactory.getILoggerFactory());
-        log = (Logger) LoggerFactory.getLogger(loggerName);
+        Logger log = (Logger) LoggerFactory.getLogger(loggerName);
         log.setLevel(Level.ALL);
         log.addAppender(logAppender);
-    }
-
-    /**
-     * @return logger that generates the logs
-     */
-    public Logger getLogger() {
-        return log;
     }
 
     /**
