+++ apiml-utility/src/test/java/org/zowe/apiml/product/logging/LogMessageTrackerTest.java
@@ -15,6 +15,7 @@ import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.slf4j.LoggerFactory;
 
@@ -35,78 +36,138 @@ class LogMessageTrackerTest {
     private final LogMessageTracker logMessageTracker = new LogMessageTracker(this.getClass());
     private final Logger log = (Logger) LoggerFactory.getLogger(this.getClass());
 
-    @BeforeEach
-    void setup() {
-        logMessageTracker.startTracking();
-        log.trace(LOG_MESSAGE);
-        log.debug(LOG_MESSAGE);
-        log.info(LOG_MESSAGE);
-        log.warn(LOG_MESSAGE);
-        log.error(LOG_MESSAGE);
+    @Nested
+    class WithoutTurboFilters {
+        @BeforeEach
+        void setup() {
+            log.getLoggerContext().resetTurboFilterList();
+            logMessageTracker.startTracking();
+            log.trace(LOG_MESSAGE);
+            log.debug(LOG_MESSAGE);
+            log.info(LOG_MESSAGE);
+            log.warn(LOG_MESSAGE);
+            log.error(LOG_MESSAGE);
+        }
+
+        @AfterEach
+        void cleanUp() {
+            logMessageTracker.stopTracking();
+        }
+
+        @Test
+        void givenLogsAndLevelFilter_whenSearch_thenFindLogs() {
+            assertEquals(5, logMessageTracker.search(LOG_MESSAGE).size());
+            assertEquals(1, logMessageTracker.search(LOG_MESSAGE, Level.INFO).size());
+
+            assertEquals(5, logMessageTracker.search(MESSAGE_REGEX).size());
+            assertEquals(1, logMessageTracker.search(MESSAGE_REGEX, Level.INFO).size());
+        }
+
+        @Test
+        void givenLogsAndLevelFilter_whenSearch_thenFindNoLogs() {
+            assertEquals(0, logMessageTracker.search(NOT_LOGGED_MESSAGE).size());
+            assertEquals(0, logMessageTracker.search(LOG_MESSAGE, Level.ALL).size());
+
+            assertEquals(0, logMessageTracker.search(NOT_MESSAGE_REGEX).size());
+            assertEquals(0, logMessageTracker.search(MESSAGE_REGEX, Level.ALL).size());
+        }
+
+        @Test
+        void givenLogsAndLevelFilter_whenCheckExist_thenLogsExist() {
+            assertTrue(logMessageTracker.contains(LOG_MESSAGE));
+            assertTrue(logMessageTracker.contains(LOG_MESSAGE, Level.TRACE));
+
+            assertTrue(logMessageTracker.contains(MESSAGE_REGEX));
+            assertTrue(logMessageTracker.contains(MESSAGE_REGEX, Level.TRACE));
+        }
+
+        @Test
+        void givenLogsAndLevelFilter_whenCheckExist_thenLogsDontExist() {
+            assertFalse(logMessageTracker.contains(NOT_LOGGED_MESSAGE));
+            assertFalse(logMessageTracker.contains(LOG_MESSAGE, Level.ALL));
+
+            assertFalse(logMessageTracker.contains(NOT_MESSAGE_REGEX));
+            assertFalse(logMessageTracker.contains(MESSAGE_REGEX, Level.ALL));
+        }
+
+        @Test
+        void givenLogs_whenGetLogs_thenAllLogsFound() {
+            assertEquals(5, logMessageTracker.countEvents());
+
+            List<ILoggingEvent> logEvents = logMessageTracker.getAllLoggedEvents();
+            logEvents.forEach(event -> assertEquals(LOG_MESSAGE, event.getFormattedMessage()));
+
+            assertEquals(1, logMessageTracker.getAllLoggedEventsWithLevel(Level.WARN).size());
+        }
+
+        @Test
+        void givenFormattedLog_whenLookForLog_thenFindFormattedLog() {
+            log.info("This is a {} log message.", "formatted");
+            assertEquals(6, logMessageTracker.countEvents());
+            assertTrue(logMessageTracker.contains("This is a formatted log message."));
+            assertEquals(1, logMessageTracker.search("This is a formatted log message.").size());
+        }
+
+        @Test
+        void givenLogTracker_whenCleanUp_thenLogsReleasedFromTracking() {
+            // Test LogMessageTracker doesn't keep logs after running LogMessageTracker.clear in @AfterEach method
+            assertEquals(5, logMessageTracker.countEvents());
+        }
     }
 
-    @AfterEach
-    void cleanUp() {
-        logMessageTracker.stopTracking();
-    }
+    @Nested
+    class WithDuplicateMessageTurboFilter {
+        ApimlDuplicateMessagesFilter duplicateMessagesFilter = new ApimlDuplicateMessagesFilter();
 
-    @Test
-    void givenLogsAndLevelFilter_whenSearch_thenFindLogs() {
-        assertEquals(5, logMessageTracker.search(LOG_MESSAGE).size());
-        assertEquals(1, logMessageTracker.search(LOG_MESSAGE, Level.INFO).size());
+        @BeforeEach
+        void setup() {
+            duplicateMessagesFilter.setAllowedRepetitions(0);
+            duplicateMessagesFilter.start();
 
-        assertEquals(5, logMessageTracker.search(MESSAGE_REGEX).size());
-        assertEquals(1, logMessageTracker.search(MESSAGE_REGEX, Level.INFO).size());
-    }
+            log.getLoggerContext().addTurboFilter(duplicateMessagesFilter);
+            logMessageTracker.startTracking();
+        }
 
-    @Test
-    void givenLogsAndLevelFilter_whenSearch_thenFindNoLogs() {
-        assertEquals(0, logMessageTracker.search(NOT_LOGGED_MESSAGE).size());
-        assertEquals(0, logMessageTracker.search(LOG_MESSAGE, Level.ALL).size());
+        @AfterEach
+        void cleanUp() {
+            logMessageTracker.stopTracking();
 
-        assertEquals(0, logMessageTracker.search(NOT_MESSAGE_REGEX).size());
-        assertEquals(0, logMessageTracker.search(MESSAGE_REGEX, Level.ALL).size());
-    }
+            // Stop all registered TurboFilters and then clear the list
+            log.getLoggerContext().resetTurboFilterList();
+        }
 
-    @Test
-    void givenLogsAndLevelFilter_whenCheckExist_thenLogsExist() {
-        assertTrue(logMessageTracker.contains(LOG_MESSAGE));
-        assertTrue(logMessageTracker.contains(LOG_MESSAGE, Level.TRACE));
+        @Test
+        void whenLoggingMessage_testApimlDuplicateMessagesFilterWorksAsExpected() {
+            RuntimeException exception = new RuntimeException("my exception");
 
-        assertTrue(logMessageTracker.contains(MESSAGE_REGEX));
-        assertTrue(logMessageTracker.contains(MESSAGE_REGEX, Level.TRACE));
-    }
+            // formattedMessage = "Message"
+            log.info("Message"); // NEUTRAL
 
-    @Test
-    void givenLogsAndLevelFilter_whenCheckExist_thenLogsDontExist() {
-        assertFalse(logMessageTracker.contains(NOT_LOGGED_MESSAGE));
-        assertFalse(logMessageTracker.contains(LOG_MESSAGE, Level.ALL));
+            // formattedMessage = "Message"
+            log.info("Message", new Object[]{"argument"}); // DENY
 
-        assertFalse(logMessageTracker.contains(NOT_MESSAGE_REGEX));
-        assertFalse(logMessageTracker.contains(MESSAGE_REGEX, Level.ALL));
-    }
+            // formattedMessage = "Message 1"
+            log.warn(String.format("Message %s", 1)); // NEUTRAL
 
-    @Test
-    void givenLogs_whenGetLogs_thenAllLogsFound() {
-        assertEquals(5, logMessageTracker.countEvents());
+            // formattedMessage = "Message 1"
+            log.warn("Message {}", new Object[]{1}); // DENY
 
-        List<ILoggingEvent> logEvents = logMessageTracker.getAllLoggedEvents();
-        logEvents.forEach(event -> assertEquals(LOG_MESSAGE, event.getFormattedMessage()));
+            // formattedMessage = "Message 1\njava.lang.RuntimeException: my exception\n...."
+            log.warn("Message 1", exception); // NEUTRAL
 
-        assertEquals(1, logMessageTracker.getAllLoggedEventsWithLevel(Level.WARN).size());
-    }
+            // formattedMessage = "Message 1\njava.lang.RuntimeException: my exception\n...."
+            log.error("Message {}", new Object[]{1, exception}); // DENY
 
-    @Test
-    void givenFormattedLog_whenLookForLog_thenFindFormattedLog() {
-        log.info("This is a {} log message.", "formatted");
-        assertEquals(6, logMessageTracker.countEvents());
-        assertTrue(logMessageTracker.contains("This is a formatted log message."));
-        assertEquals(1, logMessageTracker.search("This is a formatted log message.").size());
-    }
+            // formattedMessage = "Message [1]\njava.lang.RuntimeException: my exception\n...."
+            log.error("Message {}", new Object[]{1}, exception); // NEUTRAL
 
-    @Test
-    void givenLogTracker_whenCleanUp_thenLogsReleasedFromTracking() {
-        // Test LogMessageTracker doesn't keep logs after running LogMessageTracker.clear in @AfterEach method
-        assertEquals(5, logMessageTracker.countEvents());
+            // formattedMessage = "Message [1, java.lang.RuntimeException: Exception inside formatted message] {}\njava.lang.RuntimeException: my exception\n...."
+            log.error("Message {} {}", new Object[]{1, new RuntimeException("Exception inside formatted message")}, exception); // NEUTRAL
+
+            assertEquals(5, logMessageTracker.getAllLoggedEvents().size()); // Only messages passing the DuplicateMessageFilter are logged
+            assertEquals(1, logMessageTracker.getAllLoggedEventsWithLevel(Level.INFO).size());
+            assertEquals(2, logMessageTracker.getAllLoggedEventsWithLevel(Level.WARN).size());
+            assertEquals(2, logMessageTracker.getAllLoggedEventsWithLevel(Level.ERROR).size());
+        }
     }
 }
